Move tests.test_legacy.SoftDeleteTestCase to tests.mongo.test_documents.SoftDeleteDocumentTestCase
