+++ tests/test_legacy.py
@@ -43,99 +43,31 @@ app.config.update(
 
 db = MongoEngine(app)
 
+
 class Phone(db.Document):
     phone = PhoneField()
     strict_phone = PhoneField(strict=True)
 
+
 class Location(db.Document):
     timezone = TimezoneField()
+
 
 class TrimmedFields(db.Document):
     name = TrimmedStringField(required=True)
     comment = TrimmedStringField()
 
+
 class Secret(db.Document):
     password = EncryptedStringField(aes_generate_key())
+
 
 class Book(db.Document):
     pass
 
+
 class Author(db.Document):
     books = SafeReferenceListField(ReferenceField(Book))
-
-
-class SoftDeleteTestCase(unittest.TestCase):
-    class Person(DocumentBase, RandomPKDocument, SoftDeleteDocument):
-        name = TrimmedStringField()
-
-        meta = {
-            'allow_inheritance': True,
-        }
-
-    class Programmer(Person):
-        language = TrimmedStringField()
-
-    def setUp(self):
-        self.Person.drop_collection()
-        self.Programmer.drop_collection()
-
-    def test_default_is_deleted(self):
-        """Make sure is_deleted is never null."""
-        s = self.Person.objects.create(name='Steve')
-        self.assertEqual(s.reload()._db_data['is_deleted'], False)
-
-        def _bad_update():
-            s.update(set__is_deleted=None)
-        self.assertRaises(ValidationError, _bad_update)
-
-    def test_queryset_manager(self):
-        a = self.Person.objects.create(name='Anthony')
-
-        # test all the ways to filter/aggregate counts
-        self.assertEqual(len(self.Person.objects.all()), 1)
-        self.assertEqual(self.Person.objects.all().count(), 1)
-        self.assertEqual(self.Person.objects.filter(name='Anthony').count(), 1)
-        self.assertEqual(self.Person.objects.count(), 1)
-
-        a.delete()
-        self.assertEqual(len(self.Person.objects.all()), 0)
-        self.assertEqual(self.Person.objects.all().count(), 0)
-        self.assertEqual(self.Person.objects.filter(name='Anthony').count(), 0)
-        self.assertEqual(self.Person.objects.count(), 0)
-
-        self.assertEqual(len(self.Person.objects.filter(name='Anthony')), 0)
-        a.is_deleted = False
-        a.save()
-        self.assertEqual(len(self.Person.objects.filter(name='Anthony')), 1)
-
-        b = self.Programmer.objects.create(name='Thomas', language='python.net')
-        self.assertEqual(len(self.Programmer.objects.all()), 1)
-        b.delete()
-        self.assertEqual(len(self.Programmer.objects.all()), 0)
-
-        self.assertEqual(len(self.Programmer.objects.filter(name='Thomas')), 0)
-        b.is_deleted = False
-        b.save()
-        self.assertEqual(len(self.Programmer.objects.filter(name='Thomas')), 1)
-
-    def test_date_updated(self):
-        a = self.Person.objects.create(name='Anthony')
-        a.reload()
-        last_date_updated = a.date_updated
-
-        time.sleep(0.001)  # make sure some time passes between the updates
-        a.update(set__name='Tony')
-        a.reload()
-
-        self.assertTrue(a.date_updated > last_date_updated)
-        last_date_updated = a.date_updated
-
-        time.sleep(0.001)  # make sure some time passes between the updates
-        a.delete()
-        a.reload()
-
-        self.assertTrue(a.date_updated > last_date_updated)
-        self.assertEqual(a.is_deleted, True)
 
 
 class FieldTestCase(unittest.TestCase):
@@ -172,7 +104,6 @@ class FieldTestCase(unittest.TestCase):
         assert phone.phone == '+6594772797'
 
         phone.save()
-
 
     def test_timezone_field(self):
         location = Location()
